third_party/WebKit/LayoutTests/fast/canvas/canvas-path-addpath-expected.txt

Test addPath() method.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test addPath() with transform as identity matrix.
PASS imgdata[4] is 255
PASS imgdata[5] is 255
PASS imgdata[6] is 0
PASS imgdata[7] is 255

Test addPath() with transform as translate(100, -100).
PASS imgdata[4] is 255
PASS imgdata[5] is 255
PASS imgdata[6] is 0
PASS imgdata[7] is 255

Test addPath() with non-invertible transform.
PASS imgdata[4] is not 255
PASS imgdata[5] is not 255
PASS imgdata[6] is 0
PASS imgdata[7] is not 255

Test addPath() with transform as null or invalid type.
PASS imgdata[4] is 255
PASS imgdata[5] is 0
PASS imgdata[6] is 0
PASS imgdata[7] is 255

Test addPath() with transform omitted.
PASS imgdata[0] is 255
PASS imgdata[1] is 0
PASS imgdata[2] is 0
PASS imgdata[3] is 255

Test addPath() with path as null and invalid type
PASS p7.addPath(null, m) threw exception TypeError: Failed to execute 'addPath' on 'Path2D': The 1st argument provided is either null, or an invalid Path object..
PASS p7.addPath([], m) threw exception TypeError: Failed to execute 'addPath' on 'Path2D': The 1st argument provided is either null, or an invalid Path object..
PASS p7.addPath({}, m) threw exception TypeError: Failed to execute 'addPath' on 'Path2D': The 1st argument provided is either null, or an invalid Path object..

PASS successfullyParsed is true

TEST COMPLETE
